+++ classifai-frontend/Views/Results/PrimaryValue.swift
@@ -8,7 +8,7 @@
 import SwiftUI
 
 struct PrimaryValue: View {
-    var value: Int
+    var value: Double
     
     var body: some View {
         Text(String(value) + "%")
